MineSwepper
This is a minesweeper game made in pygames, the images in it are picked up from the internet (not owned by me). The Code however is written by me. 
those .pyc files are all that come with pygames, 
PS:- That app.exe file with Floppy Icon (that Icon is the default Icon) is the executable version of the game.